+++ tests/Pest.php
@@ -2,6 +2,6 @@
 
 declare(strict_types=1);
 
-use Simtabi\Laranail\PasswordStrength\Tests\TestCase;
+use Simtabi\Laranail\PasswordTools\Tests\TestCase;
 
 uses(TestCase::class)->in(__DIR__);